+++ packages/dashboards/examples/src/components/docs/widget-types/table/table-docs.module.ts
@@ -40,8 +40,13 @@ import { TableWidgetInteractiveExampleComponent } from "./table-widget-interacti
 import { TableSearchDocsComponent } from "./table-widget-search-docs.component";
 import { TableWidgetSearchExampleComponent } from "./table-widget-search/table-widget-search-example.component";
 import { TableWidgetExampleComponent } from "./table-widget/table-widget-example.component";
+import { TableSelectableDocsComponent } from "./table-selectable-docs.component";
+import { TableWidgetSelectableExampleComponent } from "./table-widget-selectable/table-widget-selectable.example.component";
 import { TablePaginatorDocsComponent } from "./table-paginator-docs.component";
 import { TableWidgetPaginatorExampleComponent } from "./table-widget-paginator/table-widget-paginator-example.component";
+import { TableWidgetSelectableMultiExampleComponent } from "./table-widget-selectable/table-widget-selectable-multi/table-widget-selectable-multi.example.component";
+import { TableWidgetSelectableSingleExampleComponent } from "./table-widget-selectable/table-widget-selectable-single/table-widget-selectable-single.example.component";
+import { TableWidgetSelectableRadioExampleComponent } from "./table-widget-selectable/table-widget-selectable-radio/table-widget-selectable-radio.example.component";
 
 const routes: Routes = [
     {
@@ -83,6 +88,16 @@ const routes: Routes = [
             showThemeSwitcher: true,
         },
     },
+    {
+        path: "table-select",
+        component: TableSelectableDocsComponent,
+        data: {
+            srlc: {
+                hideIndicator: true,
+            },
+            showThemeSwitcher: true,
+        },
+    },
 ];
 
 @NgModule({
@@ -99,9 +114,14 @@ const routes: Routes = [
         TableSearchDocsComponent,
         TablePaginatorDocsComponent,
         TableWidgetPaginatorExampleComponent,
+        TableSelectableDocsComponent,
+        TableWidgetInteractiveExampleComponent,
         TableWidgetExampleComponent,
         TableWidgetSearchExampleComponent,
-        TableWidgetInteractiveExampleComponent,
+        TableWidgetSelectableExampleComponent,
+        TableWidgetSelectableMultiExampleComponent,
+        TableWidgetSelectableSingleExampleComponent,
+        TableWidgetSelectableRadioExampleComponent,
     ],
     providers: [
         {