+++ packages/dashboards/src/lib/services/types.ts
@@ -21,7 +21,12 @@
 import { StaticProvider } from "@angular/core";
 import { BehaviorSubject, ReplaySubject } from "rxjs";
 
-import { EventDefinition, IEvent, IEventDefinition } from "@nova-ui/bits";
+import {
+    EventDefinition,
+    IEvent,
+    IEventDefinition,
+    ISelection,
+} from "@nova-ui/bits";
 
 import { IInteractionPayload } from "../components/providers/interaction/interaction-handler";
 import { IDataSourceBusyPayload } from "../components/providers/types";
@@ -98,6 +103,7 @@ export const DASHBOARD_EDIT_MODE = new EventDefinition<boolean>(
 export const INTERACTION = new EventDefinition<IInteractionPayload<any>>(
     "INTERACTION"
 );
+export const SELECTION = new EventDefinition<ISelection>("SELECTION");
 export const DATA_SOURCE_INVOKED = new EventDefinition<void>(
     "DATA_SOURCE_INVOKED"
 );