+++ docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [15.0.10] 📅 2024-09-04
+
+### Added
+
+- `@nova-ui/bits` | Added `selectionConfig` input and made `selectable` deprecated in the table component.
+- `@nova-ui/dashboards` | Added `selectionConfiguration` option to the table widget configuration.
+
 ## [15.0.9] 📅 2024-08-04
 
 ### Bugfix